requirements.txt
openai
langchain
streamlit
PyPDF2
faiss-cpu
tiktoken
unstructured
chromadb
huggingface_hub
google-generativeai
python-dotenv
langchain_google_genai
streamlit-extras
youtube_transcript_api
pathlib
langchain-experimental
sqlalchemy
numpy
opencv